Remove incorrect link to examples, rename variable
Hi Alex, feel free to discard this PR but I found myself wanting to find an example of how this crate can be used and didn't find one. Since I'm not in a position to provide one, my thinking was if we replace the link pointing to the JSON crate with a TODO, maybe someone will notice and add one.

LMK your thoughts, and thanks for all your work.